Convert to compile on a 64-bit box
Fixed all compilation warning too.  Some of these were related to the
int -> long change (mostly printf format mismatches)

N.b. I actually built on an os/x box, and it's possible that some of
those changesets (coming up) are in fact needed for the 64-bit build

+++ mirella/minstall.c
@@ -1621,7 +1621,7 @@ char *size;				/* expr giving size of entity */
 
    if(size == NULL) size = "0";
    sprintf(buf,"   { \"\\%03o%s\", %d, %d, %d, %d, %s },\n",
-	   strlen(name),name,type,flag,d1,d2,size);
+	   (int)strlen(name),name,type,flag,d1,d2,size);
    addbuf(6,buf);
 }
